app: clean up backup flow
Improve backup flow and fix some issues:
- Update throwIf util to directly throw non-falsy errors
- Disable parallel tests in Playwright
- Refactor backup flow tests to be more readable and thorough
- Wait for bundler request and response during backup flow
- Check for any error messages  in spec
- Add custom Playwright matcher to validate WebAuthn credentials
- Retry key added check in WebAuthn matcher
- Use new matcher in logged in account and onboarding tests
- Clean up imports and remove unnecessary code
