+++ apps/next/pages/account/settings/backup/confirm/[cred_id].tsx
@@ -1,0 +1,28 @@
+import { HomeLayout } from 'app/features/home/layout.web'
+import { SettingsLayout } from 'app/features/account/settings/layout.web'
+import { ConfirmPasskeyScreen } from 'app/features/account/settings/backup/confirm'
+import Head from 'next/head'
+import { userProtectedGetSSP } from 'utils/userProtected'
+import type { NextPageWithLayout } from '../../../../_app'
+import { ButtonOption, TopNav } from 'app/components/TopNav'
+
+export const Page: NextPageWithLayout = () => {
+  return (
+    <>
+      <Head>
+        <title>Send | Confirm Passkey</title>
+      </Head>
+      <ConfirmPasskeyScreen />
+    </>
+  )
+}
+
+export const getServerSideProps = userProtectedGetSSP()
+
+Page.getLayout = (children) => (
+  <HomeLayout TopNav={<TopNav header="Settings" button={ButtonOption.SETTINGS} />}>
+    <SettingsLayout>{children}</SettingsLayout>
+  </HomeLayout>
+)
+
+export default Page
