+++ packages/app/utils/send-accounts/useSendAccounts.ts
@@ -2,6 +2,9 @@ import { useQuery } from '@tanstack/react-query'
 import { useSupabase } from '../supabase/useSupabase'
 import { useUser } from '../useUser'
 
+/**
+ * @deprecated use useSendAccount instead
+ */
 export function useSendAccounts() {
   const { user } = useUser()
   const supabase = useSupabase()
@@ -18,6 +21,31 @@ export function useSendAccounts() {
         // no rows
         if (error.code === 'PGRST116') {
           return []
+        }
+        throw new Error(error.message)
+      }
+      return data
+    },
+  })
+}
+
+export function useSendAccount() {
+  const { user } = useUser()
+  const supabase = useSupabase()
+
+  return useQuery({
+    queryKey: ['send_account'],
+    enabled: !!user?.id,
+    queryFn: async () => {
+      const { data, error } = await supabase
+        .from('send_accounts')
+        .select('*, send_account_credentials(*, webauthn_credentials(*))')
+        .single()
+
+      if (error) {
+        // no rows
+        if (error.code === 'PGRST116') {
+          return null
         }
         throw new Error(error.message)
       }
