+++ packages/playwright/tests/fixtures/send-accounts/index.ts
@@ -1,3 +1,5 @@
+import './matchers'
+
 export {
   test,
   expect,
