+++ packages/playwright/tests/fixtures/send-accounts/matchers/index.ts
@@ -1,0 +1,1 @@
+import './webauthn-creds'
